Unify the services into a single service
