--- core/src/main/java/ai/timefold/solver/enterprise/core/partitioned/DefaultPartitionedSearchEnterpriseService.java
@@ -1,111 +1,0 @@
-package ai.timefold.solver.enterprise.core.partitioned;
-
-import static ai.timefold.solver.core.config.partitionedsearch.PartitionedSearchPhaseConfig.ACTIVE_THREAD_COUNT_AUTO;
-import static ai.timefold.solver.core.config.partitionedsearch.PartitionedSearchPhaseConfig.ACTIVE_THREAD_COUNT_UNLIMITED;
-
-import java.util.Arrays;
-import java.util.List;
-import java.util.concurrent.ThreadFactory;
-import java.util.function.BiFunction;
-
-import ai.timefold.solver.core.config.constructionheuristic.ConstructionHeuristicPhaseConfig;
-import ai.timefold.solver.core.config.localsearch.LocalSearchPhaseConfig;
-import ai.timefold.solver.core.config.partitionedsearch.PartitionedSearchPhaseConfig;
-import ai.timefold.solver.core.config.phase.PhaseConfig;
-import ai.timefold.solver.core.config.solver.EnvironmentMode;
-import ai.timefold.solver.core.config.util.ConfigUtils;
-import ai.timefold.solver.core.enterprise.PartitionedSearchEnterpriseService;
-import ai.timefold.solver.core.impl.heuristic.HeuristicConfigPolicy;
-import ai.timefold.solver.core.impl.partitionedsearch.PartitionedSearchPhase;
-import ai.timefold.solver.core.impl.partitionedsearch.partitioner.SolutionPartitioner;
-import ai.timefold.solver.core.impl.solver.termination.Termination;
-import ai.timefold.solver.core.impl.solver.thread.ChildThreadType;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-public final class DefaultPartitionedSearchEnterpriseService implements PartitionedSearchEnterpriseService {
-
-    private static final Logger LOGGER = LoggerFactory.getLogger(DefaultPartitionedSearchEnterpriseService.class);
-
-    @Override
-    public <Solution_> PartitionedSearchPhase<Solution_> buildPartitionedSearch(int phaseIndex,
-            PartitionedSearchPhaseConfig phaseConfig, HeuristicConfigPolicy<Solution_> solverConfigPolicy,
-            Termination<Solution_> solverTermination,
-            BiFunction<HeuristicConfigPolicy<Solution_>, Termination<Solution_>, Termination<Solution_>> phaseTerminationFunction) {
-        HeuristicConfigPolicy<Solution_> phaseConfigPolicy = solverConfigPolicy.createPhaseConfigPolicy();
-        ThreadFactory threadFactory = solverConfigPolicy.buildThreadFactory(ChildThreadType.PART_THREAD);
-        Termination<Solution_> phaseTermination = phaseTerminationFunction.apply(phaseConfigPolicy, solverTermination);
-        Integer resolvedActiveThreadCount = resolveActiveThreadCount(phaseConfig.getRunnablePartThreadLimit());
-        List<PhaseConfig> phaseConfigList_ = phaseConfig.getPhaseConfigList();
-        if (ConfigUtils.isEmptyCollection(phaseConfigList_)) {
-            phaseConfigList_ = Arrays.asList(new ConstructionHeuristicPhaseConfig(), new LocalSearchPhaseConfig());
-        }
-
-        DefaultPartitionedSearchPhase.Builder<Solution_> builder =
-                new DefaultPartitionedSearchPhase.Builder<>(phaseIndex, solverConfigPolicy.getLogIndentation(),
-                        phaseTermination, buildSolutionPartitioner(phaseConfig), threadFactory, resolvedActiveThreadCount,
-                        phaseConfigList_, phaseConfigPolicy.createChildThreadConfigPolicy(ChildThreadType.PART_THREAD));
-
-        EnvironmentMode environmentMode = phaseConfigPolicy.getEnvironmentMode();
-        if (environmentMode.isNonIntrusiveFullAsserted()) {
-            builder.setAssertStepScoreFromScratch(true);
-        }
-        if (environmentMode.isIntrusiveFastAsserted()) {
-            builder.setAssertExpectedStepScore(true);
-            builder.setAssertShadowVariablesAreNotStaleAfterStep(true);
-        }
-        return builder.build();
-    }
-
-    private static <Solution_> SolutionPartitioner<Solution_>
-            buildSolutionPartitioner(PartitionedSearchPhaseConfig phaseConfig) {
-        if (phaseConfig.getSolutionPartitionerClass() != null) {
-            SolutionPartitioner<?> solutionPartitioner =
-                    ConfigUtils.newInstance(phaseConfig, "solutionPartitionerClass", phaseConfig.getSolutionPartitionerClass());
-            ConfigUtils.applyCustomProperties(solutionPartitioner, "solutionPartitionerClass",
-                    phaseConfig.getSolutionPartitionerCustomProperties(), "solutionPartitionerCustomProperties");
-            return (SolutionPartitioner<Solution_>) solutionPartitioner;
-        } else {
-            if (phaseConfig.getSolutionPartitionerCustomProperties() != null) {
-                throw new IllegalStateException(
-                        "If there is no solutionPartitionerClass (" + phaseConfig.getSolutionPartitionerClass()
-                                + "), then there can be no solutionPartitionerCustomProperties ("
-                                + phaseConfig.getSolutionPartitionerCustomProperties() + ") either.");
-            }
-            // TODO Implement generic partitioner
-            throw new UnsupportedOperationException();
-        }
-    }
-
-    static Integer resolveActiveThreadCount(String runnablePartThreadLimit) {
-        return resolveActiveThreadCount(runnablePartThreadLimit, Runtime.getRuntime().availableProcessors());
-    }
-
-    static Integer resolveActiveThreadCount(String runnablePartThreadLimit, int availableProcessorCount) {
-        Integer resolvedActiveThreadCount;
-        final boolean threadLimitNullOrAuto =
-                runnablePartThreadLimit == null || runnablePartThreadLimit.equals(ACTIVE_THREAD_COUNT_AUTO);
-        if (threadLimitNullOrAuto) {
-            // Leave one for the Operating System and 1 for the solver thread, take the rest
-            resolvedActiveThreadCount = Math.max(1, availableProcessorCount - 2);
-        } else if (runnablePartThreadLimit.equals(ACTIVE_THREAD_COUNT_UNLIMITED)) {
-            resolvedActiveThreadCount = null;
-        } else {
-            resolvedActiveThreadCount = ConfigUtils.resolvePoolSize("runnablePartThreadLimit",
-                    runnablePartThreadLimit, ACTIVE_THREAD_COUNT_AUTO, ACTIVE_THREAD_COUNT_UNLIMITED);
-            if (resolvedActiveThreadCount < 1) {
-                throw new IllegalArgumentException("The runnablePartThreadLimit (" + runnablePartThreadLimit
-                        + ") resulted in a resolvedActiveThreadCount (" + resolvedActiveThreadCount
-                        + ") that is lower than 1.");
-            }
-            if (resolvedActiveThreadCount > availableProcessorCount) {
-                LOGGER.debug("The resolvedActiveThreadCount ({}) is higher than "
-                        + "the availableProcessorCount ({}), so the JVM will "
-                        + "round-robin the CPU instead.", resolvedActiveThreadCount, availableProcessorCount);
-            }
-        }
-        return resolvedActiveThreadCount;
-    }
-
-}
