--- core/src/main/resources/META-INF/services/ai.timefold.solver.core.enterprise.NearbySelectionEnterpriseService
@@ -1,1 +1,0 @@
-ai.timefold.solver.enterprise.core.nearby.DefaultNearbySelectionEnterpriseService
