+++ core/src/main/resources/META-INF/services/ai.timefold.solver.core.enterprise.TimefoldSolverEnterpriseService
@@ -1,0 +1,1 @@
+ai.timefold.solver.enterprise.core.DefaultTimefoldSolverEnterpriseService
