+++ core/src/test/java/ai/timefold/solver/enterprise/core/DefaultTimefoldSolverEnterpriseServiceTest.java
@@ -1,4 +1,4 @@
-package ai.timefold.solver.enterprise.core.partitioned;
+package ai.timefold.solver.enterprise.core;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
@@ -11,35 +11,36 @@ import ai.timefold.solver.core.impl.partitionedsearch.DefaultPartitionedSearchPh
 import ai.timefold.solver.core.impl.solver.recaller.BestSolutionRecaller;
 import ai.timefold.solver.core.impl.solver.termination.Termination;
 import ai.timefold.solver.core.impl.testdata.domain.TestdataSolution;
+import ai.timefold.solver.enterprise.core.partitioned.DefaultPartitionedSearchPhase;
 import ai.timefold.solver.enterprise.core.partitioned.testdata.TestdataSolutionPartitioner;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
-class DefaultPartitionedSearchEnterpriseServiceTest {
+public class DefaultTimefoldSolverEnterpriseServiceTest {
 
     @ParameterizedTest
     @CsvSource({
             "1, 1",
             "4, 2"
     })
-    void resolvedActiveThreadCountAuto(int availableCpuCount, int expectedResolvedCpuCount) {
-        assertThat(DefaultPartitionedSearchEnterpriseService.resolveActiveThreadCount(
+    void partitionedSearchResolvedActiveThreadCountAuto(int availableCpuCount, int expectedResolvedCpuCount) {
+        assertThat(DefaultTimefoldSolverEnterpriseService.resolveActiveThreadCount(
                 PartitionedSearchPhaseConfig.ACTIVE_THREAD_COUNT_AUTO,
                 availableCpuCount))
                 .isEqualTo(expectedResolvedCpuCount);
     }
 
     @Test
-    void resolveActiveThreadCountUnlimited() {
-        assertThat(DefaultPartitionedSearchEnterpriseService
+    void partitionedSearchResolveActiveThreadCountUnlimited() {
+        assertThat(DefaultTimefoldSolverEnterpriseService
                 .resolveActiveThreadCount(PartitionedSearchPhaseConfig.ACTIVE_THREAD_COUNT_UNLIMITED))
                 .isNull();
     }
 
     @Test
-    void assertionsForNonIntrusiveFullAssertMode() {
+    void partitionedSearchAssertionsForNonIntrusiveFullAssertMode() {
         DefaultPartitionedSearchPhase<TestdataSolution> partitionedSearchPhase =
                 mockEnvironmentMode(EnvironmentMode.NON_INTRUSIVE_FULL_ASSERT);
         assertThat(partitionedSearchPhase.isAssertStepScoreFromScratch()).isTrue();
@@ -48,7 +49,7 @@ class DefaultPartitionedSearchEnterpriseServiceTest {
     }
 
     @Test
-    void assertionsForIntrusiveFastAssertMode() {
+    void partitionedSearchAssertionsForIntrusiveFastAssertMode() {
         DefaultPartitionedSearchPhase<TestdataSolution> partitionedSearchPhase =
                 mockEnvironmentMode(EnvironmentMode.FAST_ASSERT);
         assertThat(partitionedSearchPhase.isAssertStepScoreFromScratch()).isFalse();
